Consolidate duplicated code lines
