+++ src/views/CommercialUpdateView.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
-import ReusableForm from '../components/ReusableFormComponent.vue';
-import { onMounted, ref } from 'vue';
+import CommercialEditComponent from '../components/CommercialEditComponent.vue';
 import ProjectService, { type CommercialItem } from '@/services/ProjectService';
 
 const props = defineProps<{
@@ -9,73 +8,6 @@ const props = defineProps<{
   buildingId: string;
   commercialId: string;
 }>();
-
-const fields: {
-  name: string;
-  label: string;
-  type: 'text' | 'textarea' | 'checkbox' | 'select';
-  options?: any[];
-  required?: boolean;
-  validations?: ((value: any) => string | null)[];
-}[] = [
-  { name: 'title', label: 'Titel', type: 'text', required: true },
-  { name: 'location', label: 'Standort', type: 'textarea', required: false },
-  {
-    name: 'commercialSpace',
-    label: 'Gewerbefläche (qm)',
-    type: 'text',
-    validations: [
-      (value: any) => (!isNaN(value) ? null : 'Muss eine Zahl sein'),
-      (value: any) => (value > 0 ? null : 'Muss größer als 0 sein'),
-    ],
-  },
-  {
-    name: 'usableSpace',
-    label: 'Nutzfläche (qm)',
-    type: 'text',
-    validations: [
-      (value: any) => (!isNaN(value) ? null : 'Muss eine Zahl sein'),
-      (value: any) => (value > 0 ? null : 'Muss größer als 0 sein'),
-    ],
-  },
-  {
-    name: 'heatingSpace',
-    label: 'Heizfläche (qm)',
-    type: 'text',
-    validations: [
-      (value: any) => (!isNaN(value) ? null : 'Muss eine Zahl sein'),
-      (value: any) => (value > 0 ? null : 'Muss größer als 0 sein'),
-    ],
-  },
-  {
-    name: 'description',
-    label: 'Beschreibung',
-    type: 'textarea',
-    validations: [
-      (value: any) => (value.length <= 500 ? null : 'Beschreibung muss 500 Zeichen oder weniger sein'),
-    ],
-  },
-];
-
-const initialCommercialData = ref({
-  title: '',
-  location: '',
-  commercialSpace: '',
-  usableSpace: '',
-  heatingSpace: '',
-  description: '',
-});
-
-onMounted( async ()=> {
-  const projectService = new ProjectService();
-  try {
-    const response = await projectService.getCommercial(props.projectId, props.propertyId, props.buildingId, props.commercialId);
-    console.log('Commercial data:', response.title);
-    initialCommercialData.value = response;
-  } catch (error) {
-    console.error('Error fetching commercial data:', error);
-  }
-});
 
 const handleCommercialSubmit = (values: Record<string, any>) => {
   const projectService = new ProjectService();
@@ -106,10 +38,12 @@ const handleCommercialCancel = () => {
 </script>
 
 <template>
-  <ReusableForm
+  <CommercialEditComponent
     headline="Gewerbeeinheit bearbeiten"
-    :fields="fields"
-    :initialValues="initialCommercialData"
+    :projectId="props.projectId"
+    :propertyId="props.propertyId"
+    :buildingId="props.buildingId"
+    :commercialId="props.commercialId"
     saveButtonText="Speichern"
     cancelButtonText="Abbrechen"
     @submit="handleCommercialSubmit"
